Express Framework and Router Initialization Importing Express:
The Express framework is imported, which is a minimal and flexible Node.js web application framework that provides a robust set of features for building web applications and APIs.
Creating a Router Instance:

An instance of an Express router is created using express.Router(). Routers in Express allow you to create modular, mountable route handlers. This means you can define a group of related routes and then use them in your main application.