POST Route for Creating a Booking Defining the POST Route:
A POST route is defined at the /booking endpoint. This route is responsible for handling requests to create a new movie booking.
Extracting Data from the Request Body:

The route handler extracts the movie, slot, and seats fields from the request body. These fields represent the details of the movie booking that the client wants to create.
Creating a New Ticket Instance:

A new instance of the Ticket model is created using the extracted data. This instance represents a new document that will be saved to the MongoDB collection.
Saving the Ticket Instance:

The save method of the Mongoose model is used to save the new ticket instance to the database. This is an asynchronous operation, hence the use of await.
Sending a Success Response:

If the save operation is successful, a JSON response is sent back to the client with a status code of 200 (OK). The response includes the saved ticket data and a success message.
Handling Errors:

If an error occurs during the save operation, it is caught in the catch block. An error message is sent back to the client with a status code of 500 (Internal Server Error), indicating that something went wrong on the server.