Creating a New Ticket Instance:
A new instance of the Ticket model is created using the extracted data. This instance represents a new document that will be saved to the MongoDB collection.
Saving the Ticket Instance:

The save method of the Mongoose model is used to save the new ticket instance to the database. This is an asynchronous operation, hence the use of await.
Sending a Success Response:

If the save operation is successful, a JSON response is sent back to the client with a status code of 200 (OK). The response includes the saved ticket data and a success message.