Importing Required Modules Express Framework:
The Express framework is imported to create the main web application instance. Express provides a robust set of features for building web applications and APIs, making it easier to handle HTTP requests, define routes, and manage middleware.
Creating an Express Application:

An instance of an Express application is created. This instance (app) will be used to define routes, middleware, and other settings for the web application.
Database Connection Module:

A custom module named dbConnection is imported. This module contains the logic to connect to a MongoDB database using Mongoose. It exports a function that, when called, establishes a connection to the database.
Ticket Model:

The Ticket model is imported from the schema.js file. This model defines the structure and behavior of the documents in the MongoDB collection for storing movie booking details.
CORS Middleware:

The cors middleware is imported to enable Cross-Origin Resource Sharing. CORS is a security feature implemented by browsers to restrict how resources on a web page can be requested from another domain outside the domain from which the resource originated. By using the cors middleware, the application can handle requests from different origins, making it suitable for APIs that are accessed by web applications hosted on different domains.