Middleware Setup Using CORS Middleware:
The cors middleware is added to the Express application using app.use(cors()). This allows the application to handle cross-origin requests, enabling clients from different origins to access the API without facing CORS issues.
JSON Parsing Middleware:

The express.json() middleware is added to the application. This middleware parses incoming requests with JSON payloads and makes the parsed data available in req.body. It is essential for handling POST and PUT requests where the client sends JSON data in the request body.