Importing Required Modules Express Framework:
The Express framework is imported to create the main web application instance. Express provides a robust set of features for building web applications and APIs, making it easier to handle HTTP requests, define routes, and manage middleware.
Creating an Express Application:

An instance of an Express application is created. This instance (app) will be used to define routes, middleware, and other settings for the web application.
Database Connection Module:

A custom module named dbConnection is imported. This module contains the logic to connect to a MongoDB database using Mongoose. It exports a function that, when called, establishes a connection to the database.