Importing Required Modules and Context React and State Management:
The code imports useState and useEffect from React. These hooks are used for managing state and side effects within functional components.
BsContext:

The custom context BsContext is imported. This context is likely defined elsewhere in the project and is used to provide and consume state and functions related to booking movie tickets.
Axios:

The Axios library is imported to handle HTTP requests. Axios simplifies making HTTP requests to a backend server.