Importing Required Modules and Context React and useContext Hook:
The React library and the useContext hook are imported. React is a JavaScript library for building user interfaces, and useContext is a hook that allows you to subscribe to React context changes.
RadioComponent:

The RadioComponent component is imported. This component is likely a custom component used to display radio buttons or selectable options for the movies.
moviesList:

The moviesList data is imported from a data file. This is an array of movie names or objects representing the movies that users can select from.
BsContext:

The BsContext context is imported. This context provides access to the state and functions related to the booking system, such as the selected movie and the function to change the selected movie.