Importing Required Modules and Context React and useContext Hook:
The React library and the useContext hook are imported. React is a JavaScript library for building user interfaces, and useContext is a hook that allows a component to subscribe to context changes.
RadioComponent:

The RadioComponent is imported. This component is likely used to display individual radio button options for selecting time slots.
slots Data:

The slots array is imported from a data file. This array contains the available time slots that users can choose from.
BsContext:

The BsContext context is imported. This context provides access to the global state and functions related to booking, such as the selected time slot and the function to change it.