Importing Required Modules and Context React and useContext Hook:
The React library and the useContext hook are imported. React is used for building the user interface, and useContext allows the component to subscribe to the context changes provided by BsContext.
Seats Data:

The seats array is imported from a data file. This array contains the list of available seat identifiers.
BsContext:

The BsContext context is imported. This context provides access to the global state and functions related to seat selection and booking.
SeatsInput Component:

The SeatsInput component is imported. This component is likely used to render individual seat options that the user can select or input.