Container for Seat Selection Container Div:
A div element with a class of SS_wrapper is used to contain the entire seat selection UI. It includes padding for spacing.
Title:

An h1 element is rendered with a class for styling, displaying the title "Select Seats". The text has a gradient background applied via Tailwind CSS classes, making it visually appealing and attention-grabbing.