Handler Functions changeSeatsHandler Function:
This function is defined to handle changes in the number of seats. When the input field for seat numbers is changed, this function updates the state for the number of seats.
It creates a new state object by spreading the current noOfSeat object and updating the value for the specific seat based on the input name and value.
This updated state is then saved to localStorage as a JSON string to persist the seat selection across page reloads.